pyproject.toml
[tool.poetry]
name = "ImageManip"
version = "0.0.1"
description = "Image Manipulation using Pillow"
authors = ["null2264"]
license = "Public Domain"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
